Do not request container size when listing all containers
Requesting container size in ListContainers is an expensive operation.
The list-size operation is blocking the docker daemon from doing other
container list operations (i.e. docker ps`). Even worse, the
network-daemon is also blocked from attaching weave to new containers
during this time. On my setup the list call with size takes over 5
minutes, rendering the entire setup practically useless.

There is absolutely no need for the network-daemon to know the size of
the containers, I do not know why it was set to true in the first place
but it should most definitely not be requested.